+++ packages/components/src/Shared/Entity/Miscellaneous/LogBookRenderer/index.tsx
@@ -306,7 +306,7 @@ function InternalLogBookRenderer({
             : undefined;
         return (
           <>
-            <span className="triggered-by">{localize("logbook.triggered_by_state_of")}</span>
+            <span className="triggered-by">{localize("triggered_by_state_of")}</span>
             {_renderEntity(item.context_entity_id, item.context_entity_id_name)}
             <span>
               {historicStateObj
@@ -320,7 +320,7 @@ function InternalLogBookRenderer({
       if (item.context_event_type === "call_service") {
         return (
           <>
-            <span className="triggered-by">{localize("logbook.triggered_by_action")}</span>
+            <span className="triggered-by">{localize("triggered_by_action")}</span>
             {item.context_domain && item.context_service ? (
               <span className="service-trigger-details">
                 {item.context_domain}:{" "}
@@ -341,9 +341,7 @@ function InternalLogBookRenderer({
         return (
           <>
             <span className="triggered-by">
-              {localize(
-                item.context_event_type === "automation_triggered" ? "logbook.triggered_by_automation" : "logbook.triggered_by_script",
-              )}
+              {localize(item.context_event_type === "automation_triggered" ? "triggered_by_automation" : "triggered_by_script")}
             </span>
             {_renderEntity(item.context_entity_id, item.context_entity_id_name)}
             {item.context_message ? _formatMessageWithPossibleEntity(contextTriggerSource, seenEntityIds, undefined) : null}
@@ -463,7 +461,7 @@ function InternalLogBookRenderer({
       {...rest}
     >
       {!logs.length ? (
-        <span className="no-entries">{localize("logbook.entries_not_found")}</span>
+        <span className="no-entries">{localize("no_activity_found")}</span>
       ) : (
         <Header>
           {!hideHeader && (